refactor(trello_01): change component file names
