+++ src/scripts/components/_header.js
