+++ src/scripts/index.js
@@ -1,3 +1,3 @@
-import {displayCurrentTime} from "./components/header.js";
+import {displayCurrentTime} from "./components/_header.js";
 
 setInterval(displayCurrentTime, 10);
